pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.12"
dependencies = [
    "crispy-bootstrap5>=2025.4",
    "dj-database-url>=2.3.0",
    "django>=5.2",
    "django-bootstrap5>=25.1",
    "django-crispy-forms>=2.4",
    "django-filter>=25.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "rollbar>=1.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["task_manager"]
